# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.User:
    properties:
      cityID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      photoURL:
        type: string
      role:
        $ref: '#/definitions/entity.UserRole'
      tgID:
        type: integer
      tonWallet:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  entity.UserPreferences:
    properties:
      id:
        type: string
      maxDistanceKm:
        type: integer
      maxPrice:
        type: number
      preferredCategoryID:
        type: string
      userID:
        type: string
    type: object
  entity.UserRole:
    enum:
    - master
    - client
    type: string
    x-enum-varnames:
    - UserRoleMaster
    - UserRoleClient
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a beauty services platform API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BeautyTON API
  version: "1.0"
paths:
...
  /user_preferences:
    post:
      consumes:
      - application/json
      description: Create new user preferences with the input payload
      parameters:
      - description: Create user preferences
        in: body
        name: preferences
        required: true
        schema:
          $ref: '#/definitions/entity.UserPreferences'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserPreferences'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new user preferences
      tags:
      - user-preferences
  /user_preferences/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user preferences by user preferences ID
      parameters:
      - description: User Preferences ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user preferences
      tags:
      - user-preferences
    get:
      consumes:
      - application/json
      description: Get user preferences details by user preferences ID
      parameters:
      - description: User Preferences ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserPreferences'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user preferences by ID
      tags:
      - user-preferences
    put:
      consumes:
      - application/json
      description: Update user preferences details by user preferences ID
      parameters:
      - description: User Preferences ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user preferences
        in: body
        name: preferences
        required: true
        schema:
          $ref: '#/definitions/entity.UserPreferences'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserPreferences'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user preferences
      tags:
      - user-preferences
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: Create user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - users
  /users/{id}/photo:
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Photo file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload user photo
      tags:
      - users
swagger: "2.0"